+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## main
 
+* Add `force_functional_colors` option to convert colors to functional. This change includes a deprecation warning in non-production environments that warns about non functional color usage.
+
+    *Manuel Puyol*
+
 * Promote `DetailsComponent`, `HeadingComponent`, `TextComponent`, `TimelineItemComponent`, and
   `PopoverComponent` to beta status.